added speechsynthesis api but not working
